+++ src/gflow/branches/GFlowHotFixBranch.ts
@@ -2,8 +2,6 @@ import { GitFlowBranch } from '../../api/branches/GitFlowBranch';
 import { GitFlowNodeProject } from '../../tools/GitFlowNodeProject';
 import { ProjectConfig } from '../../configs/ProjectConfig';
 import { GFlowBranch } from './GFlowBranch';
-import { Utils } from '../../tools/Utils';
-
 /**
  * This class extending a hotfix branch with some helpful functionality.
  */
@@ -49,8 +47,9 @@ export class GFlowHotFixBranch extends GFlowBranch {
     const version = await this.getVersion(project, name);
     const branchName = await this.generateBranchNameFromConfig(version);
     await project.checkoutBranch(branchName);
-    const changelogConfig = Utils.deriveChangelogConfig(this.projectConfig);
-    await project.updateChangelog(changelogConfig);
+    if (this.projectConfig?.changelog) {
+      await project.updateChangelog(this.projectConfig.changelog);
+    }
     await project.commitChanges(false);
     await super.finish(version, msg ?? version);
   }