+++ src/gflow/branches/GFlowReleaseBranch.ts
@@ -2,7 +2,6 @@ import { GitFlowBranch } from '../../api/branches/GitFlowBranch';
 import { GitFlowNodeProject } from '../../tools/GitFlowNodeProject';
 import { ProjectConfig } from '../../configs/ProjectConfig';
 import { GFlowBranch } from './GFlowBranch';
-import { Utils } from '../../tools/Utils';
 
 /**
  * This class extending a release branch with some helpful functionality.
@@ -34,8 +33,9 @@ export class GFlowReleaseBranch extends GFlowBranch {
     const branchName = await super.start(version, base);
     const project = new GitFlowNodeProject(this.projectConfig);
     await project.writeVersion(version);
-    const changelogConfig = Utils.deriveChangelogConfig(this.projectConfig);
-    await project.updateChangelog(changelogConfig, version);
+    if (this.projectConfig?.changelog) {
+      await project.updateChangelog(this.projectConfig.changelog);
+    }
     await project.commitChanges();
     return branchName;
   }